pyproject.toml
[project]
name = "deepgram-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepgram-sdk>=3.11.0",
    "mcp[cli]>=1.7.1",
]
